+++ src/views/CountryView.vue
@@ -18,29 +18,29 @@
           </RouterLink>
         </button>
       </div>
+      
+      <!-- Corrected AudioPlayer integration -->
       <div>
-        <audio controls class="styled-audio-player" v-if="audioUrl" ref="audioPlayer">
-          <source :src="audioUrl" type="audio/mpeg" />
-          Your browser does not support the audio element.
-        </audio>
+        <AudioPlayer :src="audioUrl"/>
       </div>
     </div>
   </div>
 </template>
 
 <script setup>
-import TopNav from '@/components/TopNav.vue';
 import { ref, onMounted } from 'vue';
 import { useRoute } from 'vue-router';
 import { getStorage, ref as storageRef, getDownloadURL } from 'firebase/storage';
 import { getFirestore, doc, getDoc } from 'firebase/firestore';
 import { db } from '@/firebase';  // Assuming your Firebase config is exported from this file
 
+import AudioPlayer from '@/components/AudioPlayer.vue';
+import TopNav from '@/components/TopNav.vue';
+
 const route = useRoute();
 const country = ref({ name: '', description: '', audioURL: '' });
 const imageUrl = ref('');
 const audioUrl = ref('');
-const audioPlayer = ref(null);
 
 const loadImage = async (imagePath) => {
   try {
@@ -62,29 +62,13 @@ const fetchCountryData = async (id) => {
       country.value = docSnap.data();
       const imagePath = `images/${country.value.name}.png`;
       await loadImage(imagePath);
-      audioUrl.value = country.value.audioURL;
-      console.log("Audio URL:", audioUrl.value);  // Log audio URL for debugging
+      audioUrl.value = country.value.audioUrl;
+      console.log("Audio URL:", audioUrl.value);
     } else {
       console.log("No such document!");
     }
   } catch (error) {
     console.error("Error fetching document:", error);
-  }
-};
-
-// Play audio
-const playAudio = () => {
-  if (audioPlayer.value) {
-    audioPlayer.value.play()
-      .then(() => {
-        console.log("Audio playback started");
-      })
-      .catch(error => {
-        console.error("Error playing audio:", error);
-        alert("Audio playback failed. Please check console for details.");
-      });
-  } else {
-    console.error("Audio player is not defined");
   }
 };
